test: split MountFs contract suite
