+++ crates/kernel/tests/mountfs/stream_and_union_contract.inc.rs
@@ -1,0 +1,488 @@
+
+/// A backing server with one stream file `out` (blocks at the live edge) under a
+/// directory root, used to prove a parked tail does not freeze the namespace.
+struct StreamFs {
+    stream: Stream,
+    fids: tokio::sync::Mutex<HashMap<Fid, bool>>, // fid -> is the `out` stream
+}
+
+impl StreamFs {
+    /// The server plus a direct handle to its stream, so a test can append to it
+    /// (unblocking a parked reader) without going back through the namespace.
+    fn new() -> (Self, Stream) {
+        let stream = Stream::new();
+        (
+            Self {
+                stream: stream.clone(),
+                fids: tokio::sync::Mutex::new(HashMap::new()),
+            },
+            stream,
+        )
+    }
+}
+
+#[async_trait::async_trait]
+impl FileServer for StreamFs {
+    async fn walk(&self, _fid: Fid, newfid: Fid, names: &[String]) -> Result<Qid, ErrorCode> {
+        let (is_stream, kind) = match names {
+            [] => (false, FileKind::Dir),
+            [n] if n == "out" => (true, FileKind::Stream),
+            _ => return Err(ErrorCode::NotFound),
+        };
+        self.fids.lock().await.insert(newfid, is_stream);
+        Ok(Qid {
+            kind,
+            version: 0,
+            path: 0,
+        })
+    }
+    async fn open(&self, _fid: Fid, _mode: OpenMode) -> Result<Qid, ErrorCode> {
+        Ok(Qid {
+            kind: FileKind::Stream,
+            version: 0,
+            path: 0,
+        })
+    }
+    async fn read(&self, fid: Fid, offset: Offset, count: u32) -> Result<Vec<u8>, ErrorCode> {
+        if *self.fids.lock().await.get(&fid).unwrap_or(&false) {
+            Ok(self.stream.read(offset, count).await)
+        } else {
+            Err(ErrorCode::Unsupported)
+        }
+    }
+    async fn write(&self, _fid: Fid, _offset: Offset, _data: &[u8]) -> Result<u32, ErrorCode> {
+        Err(ErrorCode::Unsupported)
+    }
+    async fn stat(&self, _fid: Fid) -> Result<Stat, ErrorCode> {
+        Ok(Stat {
+            name: String::new(),
+            qid: Qid {
+                kind: FileKind::Stream,
+                version: 0,
+                path: 0,
+            },
+            length: self.stream.len().await,
+            executable: false,
+            writable: false,
+        })
+    }
+    async fn create(&self, _: Fid, _: Fid, _: &str, _: FileKind) -> Result<Qid, ErrorCode> {
+        Err(ErrorCode::Unsupported)
+    }
+    async fn remove(&self, _: Fid) -> Result<(), ErrorCode> {
+        Err(ErrorCode::Unsupported)
+    }
+    async fn clunk(&self, fid: Fid) -> Result<(), ErrorCode> {
+        self.fids.lock().await.remove(&fid);
+        Ok(())
+    }
+}
+
+#[tokio::test]
+async fn a_parked_stream_read_does_not_freeze_the_namespace() {
+    use std::time::Duration;
+
+    // /s is a stream server (blocking live edge); /data is a MemFs. A read parked at
+    // /s/out must not hold the MountFs lock, or a concurrent op deadlocks — the M2
+    // "tail output while submitting input" workflow.
+    let (streamfs, stream) = StreamFs::new();
+    let mut ns = Namespace::new();
+    ns.mount(
+        "/s",
+        InProcessTransport::new(Arc::new(streamfs)),
+        Access::ReadWrite,
+    );
+    ns.mount("/data", memfs(), Access::ReadWrite);
+    let fs = Arc::new(MountFs::new(ns));
+
+    // A tail parks at the live edge of /s/out.
+    let tailer = fs.clone();
+    let parked = tokio::spawn(async move {
+        tailer
+            .walk(Fid::ROOT, Fid(1), &["s".into(), "out".into()])
+            .await
+            .unwrap();
+        tailer.open(Fid(1), OpenMode::Read).await.unwrap();
+        tailer.read(Fid(1), 0, 64).await.unwrap()
+    });
+    tokio::time::sleep(Duration::from_millis(20)).await;
+    assert!(
+        !parked.is_finished(),
+        "the tail should be parked at the live edge"
+    );
+
+    // A concurrent op on the same MountFs must complete despite the parked read.
+    let concurrent = tokio::time::timeout(Duration::from_millis(500), async {
+        fs.walk(Fid::ROOT, Fid(2), &["data".into(), "greeting".into()])
+            .await
+            .unwrap();
+        fs.open(Fid(2), OpenMode::Read).await.unwrap();
+        fs.read(Fid(2), 0, 64).await.unwrap()
+    })
+    .await
+    .expect("a concurrent op must not be blocked by a parked stream read");
+    assert_eq!(concurrent, b"hi");
+
+    // Appending unblocks the parked tail.
+    stream.append(b"live").await;
+    let got = tokio::time::timeout(Duration::from_millis(500), parked)
+        .await
+        .unwrap()
+        .unwrap();
+    assert_eq!(got, b"live");
+}
+
+#[tokio::test]
+async fn two_wrappers_sharing_a_backing_server_do_not_collide_backing_fids() {
+    // Two MountFs over namespaces that share the SAME ProcFs transport (as child
+    // namespaces cloning /proc would). Backing fids must be unique across wrappers,
+    // or the second wrapper's walk reuses a live fid and the shared server rejects
+    // it with BadRequest.
+    let shared = InProcessTransport::new(Arc::new(ProcFs::new()));
+    let mut ns1 = Namespace::new();
+    ns1.mount("/proc", shared.clone(), Access::ReadWrite);
+    let mut ns2 = Namespace::new();
+    ns2.mount("/proc", shared.clone(), Access::ReadWrite);
+    let fs1 = MountFs::new(ns1);
+    let fs2 = MountFs::new(ns2);
+
+    // fs1 holds a live delegated fid on the shared ProcFs.
+    fs1.walk(Fid::ROOT, Fid(1), &["proc".into(), "clone".into()])
+        .await
+        .unwrap();
+    // fs2's first walk must not collide with it.
+    fs2.walk(Fid::ROOT, Fid(1), &["proc".into(), "clone".into()])
+        .await
+        .unwrap();
+}
+
+/// A read-only directory server whose root lists a fixed set of file names, so a
+/// union of two of them (with different names) proves listing merge.
+struct DirFs {
+    entries: Vec<String>,
+    fids: tokio::sync::Mutex<HashMap<Fid, bool>>, // fid -> is the root dir
+}
+
+impl DirFs {
+    fn transport(entries: &[&str]) -> InProcessTransport {
+        InProcessTransport::new(Arc::new(DirFs {
+            entries: entries.iter().map(|s| s.to_string()).collect(),
+            fids: tokio::sync::Mutex::new(HashMap::new()),
+        }))
+    }
+}
+
+#[async_trait::async_trait]
+impl FileServer for DirFs {
+    async fn walk(&self, _fid: Fid, newfid: Fid, names: &[String]) -> Result<Qid, ErrorCode> {
+        let (is_root, kind) = match names {
+            [] => (true, FileKind::Dir),
+            [n] if self.entries.contains(n) => (false, FileKind::File),
+            _ => return Err(ErrorCode::NotFound),
+        };
+        self.fids.lock().await.insert(newfid, is_root);
+        Ok(Qid {
+            kind,
+            version: 0,
+            path: 0,
+        })
+    }
+    async fn open(&self, _fid: Fid, _mode: OpenMode) -> Result<Qid, ErrorCode> {
+        Ok(Qid {
+            kind: FileKind::Dir,
+            version: 0,
+            path: 0,
+        })
+    }
+    async fn read(&self, fid: Fid, offset: Offset, count: u32) -> Result<Vec<u8>, ErrorCode> {
+        let is_root = *self.fids.lock().await.get(&fid).unwrap_or(&false);
+        let bytes = if is_root {
+            self.entries.join("\n").into_bytes()
+        } else {
+            Vec::new()
+        };
+        let start = (offset as usize).min(bytes.len());
+        let end = bytes.len().min(start + count as usize);
+        Ok(bytes[start..end].to_vec())
+    }
+    async fn write(&self, _: Fid, _: Offset, _: &[u8]) -> Result<u32, ErrorCode> {
+        Err(ErrorCode::Unsupported)
+    }
+    async fn stat(&self, _fid: Fid) -> Result<Stat, ErrorCode> {
+        Ok(Stat {
+            name: String::new(),
+            qid: Qid {
+                kind: FileKind::Dir,
+                version: 0,
+                path: 0,
+            },
+            length: 0,
+            executable: false,
+            writable: false,
+        })
+    }
+    async fn create(&self, _: Fid, _: Fid, _: &str, _: FileKind) -> Result<Qid, ErrorCode> {
+        Err(ErrorCode::Unsupported)
+    }
+    async fn remove(&self, _: Fid) -> Result<(), ErrorCode> {
+        Err(ErrorCode::Unsupported)
+    }
+    async fn clunk(&self, fid: Fid) -> Result<(), ErrorCode> {
+        self.fids.lock().await.remove(&fid);
+        Ok(())
+    }
+}
+
+#[tokio::test]
+async fn a_union_mount_lists_all_contributors_merged() {
+    // Two servers mounted at the same prefix (a union, as `/bin` is assembled by
+    // bind). Listing must merge both contributors' entries, deduplicated.
+    let mut ns = Namespace::new();
+    ns.mount("/bin", DirFs::transport(&["ls", "cat"]), Access::ReadWrite);
+    ns.mount(
+        "/bin",
+        DirFs::transport(&["grep", "cat"]),
+        Access::ReadWrite,
+    );
+    let fs = MountFs::new(ns);
+
+    let mut entries = read_lines(&fs, &["bin"], Fid(1)).await;
+    entries.sort();
+    assert_eq!(entries, vec!["cat", "grep", "ls"]);
+}
+
+/// A directory server whose reads fail with `Io`, to prove a union merge surfaces
+/// a contributor's operational failure instead of masking it as a partial listing.
+struct FailReadDirFs;
+
+#[async_trait::async_trait]
+impl FileServer for FailReadDirFs {
+    async fn walk(&self, _fid: Fid, _newfid: Fid, names: &[String]) -> Result<Qid, ErrorCode> {
+        match names {
+            [] => Ok(Qid {
+                kind: FileKind::Dir,
+                version: 0,
+                path: 0,
+            }),
+            _ => Err(ErrorCode::NotFound),
+        }
+    }
+    async fn open(&self, _fid: Fid, _mode: OpenMode) -> Result<Qid, ErrorCode> {
+        Ok(Qid {
+            kind: FileKind::Dir,
+            version: 0,
+            path: 0,
+        })
+    }
+    async fn read(&self, _: Fid, _: Offset, _: u32) -> Result<Vec<u8>, ErrorCode> {
+        Err(ErrorCode::Io)
+    }
+    async fn write(&self, _: Fid, _: Offset, _: &[u8]) -> Result<u32, ErrorCode> {
+        Err(ErrorCode::Unsupported)
+    }
+    async fn stat(&self, _fid: Fid) -> Result<Stat, ErrorCode> {
+        Ok(Stat {
+            name: String::new(),
+            qid: Qid {
+                kind: FileKind::Dir,
+                version: 0,
+                path: 0,
+            },
+            length: 0,
+            executable: false,
+            writable: false,
+        })
+    }
+    async fn create(&self, _: Fid, _: Fid, _: &str, _: FileKind) -> Result<Qid, ErrorCode> {
+        Err(ErrorCode::Unsupported)
+    }
+    async fn remove(&self, _: Fid) -> Result<(), ErrorCode> {
+        Err(ErrorCode::Unsupported)
+    }
+    async fn clunk(&self, _fid: Fid) -> Result<(), ErrorCode> {
+        Ok(())
+    }
+}
+
+#[tokio::test]
+async fn a_union_listing_surfaces_a_contributor_failure() {
+    // One contributor's read fails operationally (Io). The merge must surface it,
+    // not silently return only the other contributor's entries.
+    let mut ns = Namespace::new();
+    ns.mount("/bin", DirFs::transport(&["ls"]), Access::ReadWrite);
+    ns.mount(
+        "/bin",
+        InProcessTransport::new(Arc::new(FailReadDirFs)),
+        Access::ReadWrite,
+    );
+    let fs = MountFs::new(ns);
+
+    fs.walk(Fid::ROOT, Fid(1), &["bin".into()]).await.unwrap();
+    fs.open(Fid(1), OpenMode::Read).await.unwrap();
+    assert_eq!(fs.read(Fid(1), 0, 4096).await, Err(ErrorCode::Io));
+}
+
+#[tokio::test]
+async fn a_relative_walk_from_a_backing_file_is_rejected() {
+    // `/data/greeting` is a file. A non-empty walk descending from it must be
+    // NotDirectory, not a re-resolution that could traverse a non-directory.
+    let fs = MountFs::new(ns());
+    fs.walk(Fid::ROOT, Fid(1), &["data".into(), "greeting".into()])
+        .await
+        .unwrap();
+    assert_eq!(
+        fs.walk(Fid(1), Fid(2), &["child".into()]).await,
+        Err(ErrorCode::NotDirectory)
+    );
+}
+
+#[tokio::test]
+async fn a_backed_directory_lists_its_mount_point_children() {
+    // A broad `/` mount (MemFs: greeting/clone/submit) plus a deeper `/mnt/llm`.
+    // Listing `/` must show the backing entries AND the `mnt` mount point, or the
+    // deeper mount is reachable but invisible in its parent's listing.
+    let mut ns = Namespace::new();
+    ns.mount("/", memfs(), Access::ReadWrite);
+    ns.mount("/mnt/llm", memfs(), Access::ReadWrite);
+    let fs = MountFs::new(ns);
+
+    let mut entries = read_lines(&fs, &[], Fid(1)).await;
+    entries.sort();
+    assert_eq!(entries, vec!["clone", "greeting", "mnt", "submit"]);
+}
+
+#[tokio::test]
+async fn remove_of_root_is_refused_and_preserves_the_handle() {
+    let fs = MountFs::new(ns());
+    assert_eq!(fs.remove(Fid::ROOT).await, Err(ErrorCode::Unsupported));
+    // The root anchor survives, so the handle still resolves absolute paths.
+    fs.walk(Fid::ROOT, Fid(1), &["data".into(), "greeting".into()])
+        .await
+        .unwrap();
+}
+
+#[tokio::test]
+async fn a_synthetic_parent_resolves_under_a_broad_mount() {
+    // A broad `/` mount (a MemFs with no `mnt` entry) plus a deeper `/mnt/llm`.
+    let mut ns = Namespace::new();
+    ns.mount("/", memfs(), Access::ReadWrite);
+    ns.mount("/mnt/llm", memfs(), Access::ReadWrite);
+    let fs = MountFs::new(ns);
+
+    // Walking to /mnt one component at a time must yield the synthetic parent of
+    // /mnt/llm, not NotFound from the broad mount lacking `mnt`.
+    assert_eq!(read_lines(&fs, &["mnt"], Fid(1)).await, vec!["llm"]);
+    // And the deeper mount is still reachable through it.
+    fs.walk(
+        Fid::ROOT,
+        Fid(2),
+        &["mnt".into(), "llm".into(), "greeting".into()],
+    )
+    .await
+    .unwrap();
+    fs.open(Fid(2), OpenMode::Read).await.unwrap();
+    assert_eq!(fs.read(Fid(2), 0, 64).await.unwrap(), b"hi");
+}
+
+#[tokio::test]
+async fn a_clunked_fid_is_released() {
+    let fs = MountFs::new(ns());
+    fs.walk(Fid::ROOT, Fid(1), &["data".into(), "greeting".into()])
+        .await
+        .unwrap();
+    fs.clunk(Fid(1)).await.unwrap();
+    // The fid is free again: walking onto it succeeds rather than BadRequest.
+    fs.walk(Fid::ROOT, Fid(1), &["data".into(), "greeting".into()])
+        .await
+        .unwrap();
+}
+
+/// A backing server whose `walk` blocks until `gate` is notified. Every other
+/// method is trivial. Used to prove `MountFs::walk` does not hold the namespace
+/// lock across the forwarded backing walk.
+struct GatedFs {
+    gate: Arc<Notify>,
+}
+
+#[async_trait::async_trait]
+impl FileServer for GatedFs {
+    async fn walk(&self, _fid: Fid, _newfid: Fid, _names: &[String]) -> Result<Qid, ErrorCode> {
+        // Block inside the forwarded walk until the test releases the gate.
+        self.gate.notified().await;
+        Ok(Qid {
+            kind: FileKind::Dir,
+            version: 0,
+            path: 0,
+        })
+    }
+    async fn open(&self, _: Fid, _: OpenMode) -> Result<Qid, ErrorCode> {
+        Ok(Qid {
+            kind: FileKind::Dir,
+            version: 0,
+            path: 0,
+        })
+    }
+    async fn read(&self, _: Fid, _: Offset, _: u32) -> Result<Vec<u8>, ErrorCode> {
+        Ok(Vec::new())
+    }
+    async fn write(&self, _: Fid, _: Offset, _: &[u8]) -> Result<u32, ErrorCode> {
+        Err(ErrorCode::NoAccess)
+    }
+    async fn stat(&self, _: Fid) -> Result<Stat, ErrorCode> {
+        Err(ErrorCode::Unsupported)
+    }
+    async fn create(&self, _: Fid, _: Fid, _: &str, _: FileKind) -> Result<Qid, ErrorCode> {
+        Err(ErrorCode::Unsupported)
+    }
+    async fn remove(&self, _: Fid) -> Result<(), ErrorCode> {
+        Err(ErrorCode::Unsupported)
+    }
+    async fn clunk(&self, _: Fid) -> Result<(), ErrorCode> {
+        Ok(())
+    }
+}
+
+/// Regression: a walk whose backing tree blocks must NOT hold the namespace
+/// lock, so a concurrent MountFs operation can still make progress. On the old
+/// code (walk held the state lock across the forwarded call) the concurrent
+/// walk deadlocked on the lock and this test timed out.
+#[tokio::test]
+async fn walk_does_not_hold_the_namespace_lock_across_the_backing_walk() {
+    let gate = Arc::new(Notify::new());
+    let mut ns = Namespace::new();
+    ns.mount(
+        "/gated",
+        InProcessTransport::new(Arc::new(GatedFs { gate: gate.clone() })),
+        Access::ReadWrite,
+    );
+    ns.mount("/data", memfs(), Access::ReadWrite);
+    let fs = Arc::new(MountFs::new(ns));
+
+    // Task 1 walks into the gated tree; its forwarded backing walk blocks.
+    let gated = {
+        let fs = fs.clone();
+        tokio::spawn(async move { fs.walk(Fid::ROOT, Fid(1), &["gated".into()]).await })
+    };
+    // Give task 1 a chance to enter the forwarded (blocked) backing walk.
+    tokio::task::yield_now().await;
+
+    // A concurrent MountFs walk needs the namespace lock. If task 1 held it
+    // across the blocked backing walk, this would deadlock. It must complete;
+    // then we release the gate so task 1 finishes too.
+    let concurrent = tokio::time::timeout(
+        std::time::Duration::from_secs(5),
+        fs.walk(Fid::ROOT, Fid(2), &["data".into()]),
+    )
+    .await
+    .expect("concurrent walk must not be blocked by the gated walk")
+    .expect("walk /data");
+    assert_eq!(concurrent.kind, FileKind::Dir);
+
+    gate.notify_one();
+    let gated = tokio::time::timeout(std::time::Duration::from_secs(5), gated)
+        .await
+        .expect("gated walk should finish after the gate is released")
+        .expect("join gated task")
+        .expect("gated walk");
+    assert_eq!(gated.kind, FileKind::Dir);
+}
